news-curator/requirements.txt
requests>=2.32.5
openai>=1.104.2
python-telegram-bot==21.5
python-dotenv==1.0.0
